anisha pulling madhura's java code

+++ TimeSheetManagement/reference_service/src/main/java/com/feuji/referenceservice/repository/CommonReferenceTypeRepo.java
@@ -5,10 +5,9 @@ import org.springframework.data.jpa.repository.Query;
 
 import com.feuji.referenceservice.entity.CommonReferenceTypeEntity;
 
-public interface CommonReferenceTypeRepo extends JpaRepository<CommonReferenceTypeEntity,Long>{
+public interface CommonReferenceTypeRepo extends JpaRepository<CommonReferenceTypeEntity,Integer>{
 
-	
-	
 	@Query(value = "select reference_type_id from common_reference_type where reference_type_name=?",nativeQuery = true)
 	public CommonReferenceTypeEntity getByTypeName(String typeName);
+	
 }
